Moved aboutMe, projects from pages to components folder
