+++ components/sections/projects.tsx
@@ -1,8 +1,8 @@
-import ProjectCard from "../components/projectCard"
+import ProjectCard from "../projectCard"
 
 const Projects = () => {
     return (
-        <section id="projects" className="h-screen">
+        <section id="projects" className="min-h-screen flex flex-col">
             <h1 className="text-3xl mx-1 my-5 md:text-5xl lg:mx-7 lg:my-7">Projects</h1>
             <div className="h-content grid grid-cols-1 sm:grid-cols-3 sm:gap-12 sm:mx-12 sm:my-12">
                 <ProjectCard
