+++ pages/index.tsx
@@ -1,14 +1,16 @@
 import Footer from '../components/footer'
 import Layout from '../components/layout'
-import Projects from './projects'
-import Home from './home'
+import Projects from '../components/sections/projects'
+import AboutMe from '../components/sections/aboutMe'
+import ContactMe from '../components/sections/contactMe'
 
 export default function Index() {
   return (
     <Layout>
       <main className="flex flex-col">
-        <Home />
+        <AboutMe />
         <Projects />
+        <ContactMe />
         <Footer />
       </main >
     </Layout>
